hallo

+++ commands/say.js
@@ -1,6 +1,8 @@
 const errors = require('../utils/errors.js')
 
 module.exports.run = async (bot, message, args)=>{
+    if(!message.author.hasPermission["MANAGE_MESSAGES"])return errors.noPerms(message, "MANAGE_MESSAGES")
+
     message.delete()
     let botmessage= args.join(" ")
     message.channel.send(botmessage)
